# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Alert": {
        "additionalProperties": false,
        "description": "An advisory raised by the conformance service.\n\n``reason_codes`` carries machine-readable justifications so an alert is\nalways explainable without re-running the detector.",
        "properties": {
          "alert_id": {
            "title": "Alert Id",
            "type": "string"
          },
          "alert_key": {
            "title": "Alert Key",
            "type": "string"
          },
          "conflict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConflictEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "conformance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConformanceEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/AlertKind"
          },
          "raised_at": {
            "format": "date-time",
            "title": "Raised At",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "schema_version": {
            "const": "1",
            "default": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "source": {
            "$ref": "#/components/schemas/DataSource",
            "default": "simulator"
          },
          "state": {
            "$ref": "#/components/schemas/AlertState"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "track_ids": {
            "items": {
              "type": "string"
            },
            "title": "Track Ids",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "alert_key",
          "kind",
          "severity",
          "state",
          "raised_at",
          "updated_at",
          "track_ids",
          "reason_codes",
          "summary"
        ],
        "title": "Alert",
        "type": "object"
      },
      "AlertKind": {
        "description": "What kind of condition was detected.",
        "enum": [
          "predicted_conflict",
          "non_conformance",
          "emergency_squawk",
          "excessive_descent"
        ],
        "title": "AlertKind",
        "type": "string"
      },
      "AlertState": {
        "description": "Alert lifecycle, used to suppress flapping.\n\nAn alert is raised once as ``NEW``, re-published as ``SUSTAINED`` while the\ncondition persists, and closed as ``CLEARED``. Consumers key on\n``alert_key`` and keep only the latest state.",
        "enum": [
          "new",
          "sustained",
          "cleared"
        ],
        "title": "AlertState",
        "type": "string"
      },
      "AlertsResponse": {
        "description": "Everything the system currently considers wrong.\n\nCleared alerts are absent rather than flagged: the contract is \"what is\nwrong now\", so a consumer cannot forget to filter them out.",
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "count",
          "alerts"
        ],
        "title": "AlertsResponse",
        "type": "object"
      },
      "ConflictEvidence": {
        "additionalProperties": false,
        "description": "Geometry behind a predicted loss of separation.\n\nDistances are the *predicted minimum* over the lookahead window, not the\ncurrent separation.",
        "properties": {
          "horizontal_standard_nm": {
            "exclusiveMinimum": 0.0,
            "title": "Horizontal Standard Nm",
            "type": "number"
          },
          "lookahead_s": {
            "exclusiveMinimum": 0.0,
            "title": "Lookahead S",
            "type": "number"
          },
          "min_horizontal_sep_nm": {
            "minimum": 0.0,
            "title": "Min Horizontal Sep Nm",
            "type": "number"
          },
          "min_vertical_sep_ft": {
            "minimum": 0.0,
            "title": "Min Vertical Sep Ft",
            "type": "number"
          },
          "time_to_cpa_s": {
            "title": "Time To Cpa S",
            "type": "number"
          },
          "vertical_standard_ft": {
            "exclusiveMinimum": 0.0,
            "title": "Vertical Standard Ft",
            "type": "number"
          }
        },
        "required": [
          "time_to_cpa_s",
          "min_horizontal_sep_nm",
          "min_vertical_sep_ft",
          "lookahead_s",
          "horizontal_standard_nm",
          "vertical_standard_ft"
        ],
        "title": "ConflictEvidence",
        "type": "object"
      },
      "ConformanceEvidence": {
        "additionalProperties": false,
        "description": "How far an aircraft diverged from where it was predicted to be.",
        "properties": {
          "baseline_error_nm": {
            "minimum": 0.0,
            "title": "Baseline Error Nm",
            "type": "number"
          },
          "error_nm": {
            "minimum": 0.0,
            "title": "Error Nm",
            "type": "number"
          },
          "horizon_s": {
            "exclusiveMinimum": 0.0,
            "title": "Horizon S",
            "type": "number"
          },
          "predicted_at": {
            "format": "date-time",
            "title": "Predicted At",
            "type": "string"
          },
          "predictor_version": {
            "title": "Predictor Version",
            "type": "string"
          },
          "threshold_nm": {
            "exclusiveMinimum": 0.0,
            "title": "Threshold Nm",
            "type": "number"
          }
        },
        "required": [
          "predicted_at",
          "horizon_s",
          "error_nm",
          "baseline_error_nm",
          "threshold_nm",
          "predictor_version"
        ],
        "title": "ConformanceEvidence",
        "type": "object"
      },
      "DataSource": {
        "description": "Provenance of a message.\n\nEvery message declares where it came from. There is currently one value; it\nexists so that a synthetic record can never be mistaken for a real one if a\nlive feed is ever added.",
        "enum": [
          "simulator"
        ],
        "title": "DataSource",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "Liveness: this process is running.",
        "properties": {
          "service": {
            "default": "api",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "HealthResponse",
        "type": "object"
      },
      "HistoryResponse": {
        "description": "Where a track has been.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/TrackPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "track_id": {
            "title": "Track Id",
            "type": "string"
          }
        },
        "required": [
          "track_id",
          "count",
          "points"
        ],
        "title": "HistoryResponse",
        "type": "object"
      },
      "ReadyResponse": {
        "description": "Readiness: dependencies are reachable.",
        "properties": {
          "postgres": {
            "title": "Postgres",
            "type": "boolean"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "redis": {
            "title": "Redis",
            "type": "boolean"
          }
        },
        "required": [
          "ready",
          "redis",
          "postgres"
        ],
        "title": "ReadyResponse",
        "type": "object"
      },
      "Severity": {
        "description": "Advisory severity. This system never issues a control instruction.",
        "enum": [
          "info",
          "advisory",
          "caution",
          "warning"
        ],
        "title": "Severity",
        "type": "string"
      },
      "TrackPoint": {
        "description": "One historical position.",
        "properties": {
          "altitude_ft": {
            "title": "Altitude Ft",
            "type": "number"
          },
          "ground_speed_kt": {
            "title": "Ground Speed Kt",
            "type": "number"
          },
          "lat": {
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "title": "Lon",
            "type": "number"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "position_uncertainty_m": {
            "title": "Position Uncertainty M",
            "type": "number"
          },
          "track_deg": {
            "title": "Track Deg",
            "type": "number"
          },
          "vertical_rate_fpm": {
            "title": "Vertical Rate Fpm",
            "type": "number"
          }
        },
        "required": [
          "observed_at",
          "lat",
          "lon",
          "altitude_ft",
          "ground_speed_kt",
          "track_deg",
          "vertical_rate_fpm",
          "position_uncertainty_m"
        ],
        "title": "TrackPoint",
        "type": "object"
      },
      "TrackState": {
        "description": "Lifecycle of a track inside the tracker.",
        "enum": [
          "initiating",
          "confirmed",
          "coasting",
          "terminated"
        ],
        "title": "TrackState",
        "type": "string"
      },
      "TrackUpdate": {
        "additionalProperties": false,
        "description": "A filtered state estimate for one aircraft.\n\nUnlike :class:`SurveillanceReport` the kinematic fields are always present:\nthe filter always has an estimate, even while coasting through a dropout.\n``position_uncertainty_m`` is how much to trust it.",
        "properties": {
          "altitude_ft": {
            "title": "Altitude Ft",
            "type": "number"
          },
          "callsign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callsign"
          },
          "ground_speed_kt": {
            "minimum": 0.0,
            "title": "Ground Speed Kt",
            "type": "number"
          },
          "icao24": {
            "pattern": "^[0-9a-f]{6}$",
            "title": "Icao24",
            "type": "string"
          },
          "innovation_nm": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Innovation Nm"
          },
          "last_report_at": {
            "format": "date-time",
            "title": "Last Report At",
            "type": "string"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "position_uncertainty_m": {
            "minimum": 0.0,
            "title": "Position Uncertainty M",
            "type": "number"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "schema_version": {
            "const": "1",
            "default": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/DataSource",
            "default": "simulator"
          },
          "squawk": {
            "anyOf": [
              {
                "pattern": "^[0-7]{4}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Squawk"
          },
          "state": {
            "$ref": "#/components/schemas/TrackState"
          },
          "track_deg": {
            "exclusiveMaximum": 360.0,
            "minimum": 0.0,
            "title": "Track Deg",
            "type": "number"
          },
          "track_id": {
            "title": "Track Id",
            "type": "string"
          },
          "turn_rate_deg_s": {
            "title": "Turn Rate Deg S",
            "type": "number"
          },
          "update_count": {
            "minimum": 0.0,
            "title": "Update Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "vertical_rate_fpm": {
            "title": "Vertical Rate Fpm",
            "type": "number"
          }
        },
        "required": [
          "track_id",
          "icao24",
          "updated_at",
          "last_report_at",
          "state",
          "lat",
          "lon",
          "altitude_ft",
          "ground_speed_kt",
          "track_deg",
          "vertical_rate_fpm",
          "turn_rate_deg_s",
          "position_uncertainty_m",
          "update_count"
        ],
        "title": "TrackUpdate",
        "type": "object"
      },
      "TracksResponse": {
        "description": "The current airspace picture.",
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "live_window_s": {
            "default": 20.0,
            "title": "Live Window S",
            "type": "number"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/TrackUpdate"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "count",
          "tracks"
        ],
        "title": "TracksResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only view of the live airspace picture and track history.\n\n**All data is synthetic.** See `docs/safety-notes.md` in the repository: this system is advisory only, is not certified, and must never be used for actual air traffic control.",
    "summary": "Advisory airspace monitoring over simulated traffic. Not an ATC system.",
    "title": "Airspace Conformance Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Liveness. Deliberately checks nothing external.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "operations"
        ]
      }
    },
    "/ready": {
      "get": {
        "description": "Readiness. Reports each dependency separately so a failure names itself.",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponse"
                }
              }
            },
            "description": "A dependency is unreachable"
          }
        },
        "summary": "Ready",
        "tags": [
          "operations"
        ]
      }
    },
    "/v1/alerts": {
      "get": {
        "description": "Active advisories, most recently updated first.\n\nRepeat `kind` to filter: `?kind=predicted_conflict&kind=emergency_squawk`.\nOmit it entirely for everything.\n\nA repeated parameter rather than a single nullable one. `AlertKind | None`\nproduced a spec claiming the parameter accepted null, which has no\nmeaningful representation in a query string and which the server then\nrejected with a 422 -- an implementation that contradicted its own\ncontract. Schemathesis found it by sending `?kind=null`. Filtering by\nseveral kinds at once is more useful anyway.\n\nThese are advisory only. Nothing here is an instruction, and the system\nnever proposes a resolution - see `docs/safety-notes.md`.",
        "operationId": "alerts_v1_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/AlertKind"
              },
              "title": "Kind",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alerts",
        "tags": [
          "airspace"
        ]
      }
    },
    "/v1/tracks": {
      "get": {
        "description": "Every track seen within the last `window_s` seconds.",
        "operationId": "tracks_v1_tracks_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_s",
            "required": false,
            "schema": {
              "default": 20.0,
              "exclusiveMinimum": 0.0,
              "maximum": 600.0,
              "title": "Window S",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tracks",
        "tags": [
          "airspace"
        ]
      }
    },
    "/v1/tracks/{track_id}/history": {
      "get": {
        "description": "Recent positions for one track, oldest first.",
        "operationId": "history_for_v1_tracks__track_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "title": "Track Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 600,
              "exclusiveMinimum": 0,
              "maximum": 5000,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No history for that track"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History For",
        "tags": [
          "airspace"
        ]
      }
    }
  }
}